Amend "extract" scope
Both the "positive" and "negative" methods had private "extract" functions, which differed only by the value of the condition they were testing for ("true" or "null"). Performance testing reveals that a shared private "extract" function scoped to the class (parallel in scope with the "positive" and "negative" methods themselves) performs marginally faster, and there is no appreciable difference in the performance of the "extract" function if the condition is passed as an argument. Consequently, neither "positive" nor "negative" need themselves to be defined inside a self-executing function.