Amend "DifferenceEngine" constructor
Privileged methods are now appended to the constructor's prototype. Performance testing indicates that declaring methods in parallel to the constructor rather than as they are assigned onto the prototype is more efficient.

A constructor without a return value instantiates fastest in IE and FF but slowest in Chrome: however, Chrome instantiates objects faster than the other browsers by a significant margin regardless of the return value so that result is irrelevant. I have chosen to return "this" although it is not the fastest because I prefer it stylistically.