Amend "order" method
The "order" method expects that the "alpha" and "omega" arrays contain
the same elements but not in the same order (so the array length should
be one or greater and both the "positive" and "negative" methods should
return empty arrays). The "extracted" array returns the elements whose
positions differ between arrays.